Refactor to handle multiple mentions of Git repo
The core of this change rewires the database so that hnposts are children of git repositories.  In the context of this project, this makes sense: we care about the repos, and 1 or more hnposts has mentioned them.

This allows us to do some interesting things, including displaying and linking to the first and last mentions of a project, and the total mentions.

I couldn't find a `knex` syntax for selecting * from just one table, e.g. `select * from a.*`. So listed out all the columns needed. Refactored shared parts of the query into a base query that could be shared, minimizing duplicated code.

Update clearOldPosts to drop both the ghprojects and hnposts. This can be removed if clearOldPosts isn't needed.

This requires a complete drop, re-add, and then re-population of the database